Completely re-factored feature extraction
Any existing dependencies on HOG specifics have been removed from ARTOS and FFLD, which are not strictly separated any more.

A new feature extraction framework provides an abstraction layer which should make it easy to implement arbitrary feature extractors. The type and parameters of the feature extractor used to learn a model will be saved with the model. That means, the file format of model files has changed, as well as the format of background statistics, but ARTOS can still read the old format.

The speed of training and detection has also been improved further.

The detection still behaves a bit ominous at the moment.